+++ lib/Context/index.d.ts
@@ -1,0 +1,4 @@
+import * as React from "react";
+import { DebugContextProps } from "./types";
+export declare const DebugContext: React.Context<DebugContextProps>;
+//# sourceMappingURL=index.d.ts.map