+++ src/context.ts
@@ -1,5 +1,0 @@
-import * as React from "react";
-
-export const DebugContext: React.Context<boolean> = React.createContext(
-  false as boolean
-);