+++ src/index.ts
@@ -1,7 +1,12 @@
-export { DebugContext } from "./context";
-export { Inset } from "./Inset";
-export { insetFactory } from "./Inset/insetFactory";
-export {
+import { DebugContext } from "./Context";
+import {
+  DebugContextProps,
+  DebugItemProps,
+  InsetDebugItemProps
+} from "./Context/types";
+import Inset from "./Inset";
+import insetFactory from "./Inset/insetFactory";
+import {
   All,
   InsetDebugOptions,
   Horizontal,
@@ -12,9 +17,37 @@ export {
   Vertical,
   InsetStyles
 } from "./Inset/insetTypes";
-export { Stack } from "./Stack";
-export { stackFactory } from "./Stack/stackFactory";
-export { StackDebugOptions, StackProps, StackStyles } from "./Stack/stackTypes";
-export { Queue } from "./Queue";
-export { queueFactory } from "./Queue/queueFactory";
-export { QueueDebugOptions, QueueProps, QueueStyles } from "./Queue/queueTypes";
+import Stack from "./Stack";
+import stackFactory from "./Stack/stackFactory";
+import { StackDebugOptions, StackProps, StackStyles } from "./Stack/stackTypes";
+import Queue from "./Queue";
+import queueFactory from "./Queue/queueFactory";
+import { QueueDebugOptions, QueueProps, QueueStyles } from "./Queue/queueTypes";
+
+export {
+  DebugContext,
+  DebugContextProps,
+  InsetDebugItemProps,
+  DebugItemProps,
+  Inset,
+  insetFactory,
+  All,
+  InsetDebugOptions,
+  Horizontal,
+  InsetProps,
+  Other,
+  PaddingPossibilities,
+  VerHor,
+  Vertical,
+  InsetStyles,
+  Stack,
+  stackFactory,
+  StackDebugOptions,
+  StackProps,
+  StackStyles,
+  Queue,
+  queueFactory,
+  QueueDebugOptions,
+  QueueProps,
+  QueueStyles
+};